td-shim-tools: rename td-shim-enroll-key as td-shim-tools
Rename td-shim-enroll-key as td-shim-tools, later td-shim-ld and
td-shim-sign-payload will be merged into this crate.
