+++ td-shim-tools/Cargo.toml
@@ -1,20 +1,21 @@
 [package]
-name = "td-shim-enroll-key"
+name = "td-shim-tools"
 version = "0.1.0"
-description = "Enroll hash of a DER encoded public key into the Configuration Firmware Volume"
+description = "A set of tools to manipulate td-shim binary"
 repository = "https://github.com/confidential-containers/td-shim"
 homepage = "https://github.com/confidential-containers"
 license = "BSD-2-Clause-Patent"
 edition = "2018"
 
-[features]
-default = ["enroller"]
-enroller = ["clap", "der", "env_logger", "log", "ring", "td-shim-ld/linker"]
+[[bin]]
+name = "td-shim-enroll-key"
+required-features = ["enroller"]
 
 [dependencies]
 r-efi = "3.2.0"
 scroll = { version = "0.10", default-features = false, features = ["derive"]}
 td-layout = { path = "../td-layout" }
+td-shim = { path = "../td-shim" }
 td-shim-ld = { path = "../td-shim-ld", default-features = false }
 uefi-pi =  { path = "../uefi-pi" }
 
@@ -23,3 +24,7 @@ der = { version = "0.4.5", features = ["oid"], optional = true }
 env_logger = { version = "0.9.0", optional = true }
 log = { version = "0.4.5", optional = true }
 ring = { git = "https://github.com/jyao1/ring.git", branch = "uefi_support", default-features = false, features = ["alloc"], optional = true }
+
+[features]
+default = ["enroller"]
+enroller = ["clap", "der", "env_logger", "log", "ring", "td-shim/secure-boot", "td-shim-ld/linker"]
