+++ td-shim-tools/src/lib.rs
@@ -1,0 +1,8 @@
+// Copyright (c) 2022 Alibaba Cloud
+//
+// SPDX-License-Identifier: BSD-2-Clause-Patent
+
+#[cfg(feature = "enroller")]
+pub mod enroller;
+#[cfg(feature = "enroller")]
+pub mod public_key;
