+++ td-shim-tools/src/public_key.rs
