+++ td-shim/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust-tdshim"
+name = "td-shim"
 version = "0.1.0"
 edition = "2018"
 
@@ -18,6 +18,7 @@ tdx-tdcall = { path = "../tdx-tdcall" }
 [dependencies]
 lazy_static = { version = "1.4.0", features = ["spin_no_std"] }
 log = "0.4.13"
+r-efi = "3.2.0"
 scroll = { version = "0.10", default-features = false, features = ["derive"] }
 uefi-pi =  { path = "../uefi-pi" }
 zerocopy = "0.6.0"
@@ -26,7 +27,6 @@ elf-loader = { path = "../elf-loader", optional = true }
 linked_list_allocator = { version = "0.9.0", optional = true }
 ring = { git = "https://github.com/jyao1/ring.git", branch = "uefi_support", default-features = false, features = ["alloc"], optional = true }
 pe-loader = { path = "../pe-loader", optional = true }
-r-efi = { version = "3.2.0", optional = true }
 spin = { version = "0.9.2", optional = true }
 td-exception =  { path = "../td-exception", features = ["tdx"], optional = true }
 td-layout = { path = "../td-layout", optional = true }
@@ -38,7 +38,6 @@ x86_64 = { version = "=0.14.6", optional = true }
 
 # secure boot
 der = {version = "0.4.5", features = ["derive", "alloc"], optional = true}
-td-shim-enroll-key = { path = "../td-shim-enroll-key", default-features = false, optional = true }
 td-shim-sign-payload = { path = "../td-shim-sign-payload", default-features = false, optional = true }
 
 # TDX
@@ -48,13 +47,12 @@ tdx-tdcall = { path = "../tdx-tdcall", optional = true }
 default = ["boot-kernel", "cet-ss", "secure-boot"]
 boot-kernel = ["td-layout/boot-kernel"]
 cet-ss = []
-secure-boot = ["der", "td-shim-enroll-key", "td-shim-sign-payload"]
+secure-boot = ["der", "td-shim-sign-payload"]
 tdx = ["tdx-tdcall", "td-exception/tdx", "td-logger/tdx", "x86"]
 main = [
     "elf-loader",
     "linked_list_allocator",
     "pe-loader",
-    "r-efi",
     "ring",
     "spin",
     "td-exception",
